Add files via upload
This is basic calculator program which performs arithmetic operations on numbers like Addition, Subtraction, Division, Exponential, Multiplication, division-Remainder. Two numbers are taken as an input and one more input to choose operator (according to need of arithmetic output). Then using if-else statements and user defined functions we calculate the arithmetic output  and then we display these results in a list so that we can get required output in a single row.